requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
boto3==1.35.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
Django==5.0.4
django-qr-code==4.0.1
djangorestframework==3.15.1
idna==3.7
OTPLessAuthSDK==0.3.3
Pillow==9.5.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
rsa==4.9
segno==1.6.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
